Removed old method and attribute plot components
