+++ src/middlePanelViews/plotter.component.test.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import { createShallow } from '@material-ui/core/test-utils';
-import MockCircularBuffer from '../../malcolm/reducer/attribute.reducer.mocks';
-import { malcolmTypes } from '../../malcolmWidgets/attributeDetails/attributeSelector/attributeSelector.component';
-import AttributePlot from './attributePlot.component';
+import MockCircularBuffer from '../malcolm/reducer/attribute.reducer.mocks';
+import { deriveStateFromProps as deriveMethodState } from './methodView/methodArchive.container';
+import { deriveStateFromProps as deriveAttributeState } from './attributeView/attributeView.container';
+import { malcolmTypes } from '../malcolmWidgets/attributeDetails/attributeSelector/attributeSelector.component';
+
+import Plotter from './plotter.component';
 
 const mockTheme = {
   palette: {
@@ -12,6 +15,64 @@ const mockTheme = {
   },
   alarmState: { warning: '#fff', error: '#ff0000', disconnected: '#0500ff' },
 };
+
+describe('MethodPlot', () => {
+  let shallow;
+  let testArchive;
+
+  beforeEach(() => {
+    shallow = createShallow({ dive: true });
+    testArchive = {
+      value: new MockCircularBuffer(5),
+      timeStamp: new MockCircularBuffer(5),
+      alarmState: new MockCircularBuffer(5),
+    };
+    testArchive.value.push({
+      runParameters: { test1: 0 },
+      returned: { test2: 1 },
+    });
+    testArchive.timeStamp.push({
+      localRunTime: new Date(0),
+      localReturnTime: new Date(4),
+    });
+    testArchive.alarmState.push(0);
+    testArchive.value.push({
+      runParameters: { test1: 0 },
+      returned: { test2: 1 },
+    });
+    testArchive.timeStamp.push({
+      localRunTime: new Date(2),
+      localReturnTime: new Date(5),
+    });
+    testArchive.alarmState.push(1);
+    testArchive.value.push({
+      runParameters: { test1: 0 },
+      returned: { test2: 1 },
+    });
+    testArchive.timeStamp.push({
+      localRunTime: new Date(3),
+      localReturnTime: new Date(6),
+    });
+    testArchive.alarmState.push(2);
+  });
+
+  it('renders correctly', () => {
+    const wrapper = shallow(
+      <Plotter
+        store={{
+          getState: () => {},
+          subscribe: () => {},
+        }}
+        attribute={testArchive}
+        openPanels={{}}
+        selectedParam={['takes', 'test1']}
+        theme={mockTheme}
+        deriveState={deriveMethodState}
+      />
+    );
+    expect(wrapper.dive()).toMatchSnapshot();
+  });
+});
 
 describe('attributePlot', () => {
   let shallow;
@@ -45,11 +106,13 @@ describe('attributePlot', () => {
     mockArchive.alarmState.push(0);
 
     const wrapper = shallow(
-      <AttributePlot
+      <Plotter
         store={mockStore}
         attribute={mockArchive}
         openPanels={{}}
         theme={mockTheme}
+        deriveState={deriveAttributeState}
+        doTick
       />
     );
     expect(wrapper.dive()).toMatchSnapshot();
@@ -68,11 +131,13 @@ describe('attributePlot', () => {
     mockArchive.alarmState.push(0);
 
     const wrapper = shallow(
-      <AttributePlot
+      <Plotter
         store={mockStore}
         attribute={mockArchive}
         openPanels={{}}
         theme={mockTheme}
+        deriveState={deriveAttributeState}
+        doTick
       />
     );
     expect(wrapper.dive()).toMatchSnapshot();
@@ -91,11 +156,13 @@ describe('attributePlot', () => {
     mockArchive.alarmState.push(0);
 
     const wrapper = shallow(
-      <AttributePlot
+      <Plotter
         store={mockStore}
         attribute={mockArchive}
         openPanels={{}}
         theme={mockTheme}
+        deriveState={deriveAttributeState}
+        doTick
       />
     );
     expect(wrapper.dive()).toMatchSnapshot();
